Fixed improper usage of iterater while erasing an element in a map
The iterator was being incremented as part of the for loop after it had been
erased. The erase method invalidates the iterator making the increment fail in
VS2012 Express Edition. The increment apparently worked using g++ 4.8.1.

The post-increment now being performed in the call to erase works because the
standard explicitly states that the increment must be completed before the call
to erase is executed.